requirements.txt
jaydebeapi==1.2.3
psycopg2-binary==2.9.9
pandas==2.0.3
openpyxl==3.1.2
xlwings==0.31.1
python-dotenv==1.0.0
pyyaml==6.0.1